+++ GRDB/DateTime.swift
@@ -1,0 +1,74 @@
+//
+// GRDB.swift
+// https://github.com/groue/GRDB.swift
+// Copyright (c) 2015 Gwendal Roué
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+
+public struct DateTime: DatabaseValueConvertible {
+    
+    // MARK: - NSDate conversion
+    //
+    // We consistently use the Swift nil to represent the database NULL: the
+    // date property is a non-optional NSDate, and the NSDate initializer is
+    // failable:
+    
+    /// The represented date
+    public let date: NSDate
+    
+    /// Creates a DateTime from an NSDate.
+    /// Returns nil if and only if the NSDate is nil.
+    public init?(_ date: NSDate?) {
+        if let date = date {
+            self.date = date
+        } else {
+            return nil
+        }
+    }
+    
+    
+    // MARK: - DatabaseValue conversion
+    //
+    // DateTime represents an NSDate as an ISO-8601 string.
+    
+    // An ISO-8601 date formatter
+    static let dateFormatter: NSDateFormatter = {
+        let formatter = NSDateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ss.SSS"
+        formatter.locale = NSLocale(localeIdentifier: "en_US_POSIX")
+        formatter.timeZone = NSTimeZone(forSecondsFromGMT: 0)
+        return formatter
+    }()
+    
+    public var databaseValue: DatabaseValue {
+        return .Text(DateTime.dateFormatter.stringFromDate(date))
+    }
+    
+    public init?(databaseValue: DatabaseValue) {
+        // Don't handle the raw DatabaseValue since GRDB built-in conversions
+        // do all the job for us:
+        if let string = String(databaseValue: databaseValue) {
+            self.init(DateTime.dateFormatter.dateFromString(string))
+        } else {
+            return nil
+        }
+    }
+}
+